Do not allow unused trailing bytes in BAM records
Previously, the .data field of BAM records could have unused bytes near the end.
This is acceptable, because the length of the actually used bytes is known from
the fixed fields of the record.
It had the advantage of preventing frequent resizes, which was slow in Julia.
However, from Julia 1.11 onwards, resizing is fast.
Promising no non-coding bytes in the BAM records have the following advantages:
* It may simplify some code in XAM
* It allows an immutable struct holding a reference to the vector to resize
  the vector, signalling a change in the variable-length data.